fixed firebase warnings and got errors to show up on login
