+++ puffpuffpets.ui/src/Components/MyNavbar/LoginModal/LoginModal.scss
@@ -1,0 +1,3 @@
+.error {
+  color: red;
+}
