+++ puffpuffpets.ui/src/Helpers/Data/fbConnection.js
@@ -1,4 +1,4 @@
-import firebase from 'firebase';
+import firebase from 'firebase/app';
 import constants from '../apiKeys.json';
 
 const firebaseApp = () => {
