+++ puffpuffpets.ui/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
-import 'bootstrap/dist/css/bootstrap.min.css';import App from './Components/App/App.js';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import App from './Components/App/App.js';
 
 ReactDOM.render(<App />, document.getElementById('root'));
